django/requirements.txt
django==3.2
python-dotenv==1.0.0
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
django-postgres-extra==2.0.8
psycopg2==2.9.7
gunicorn==21.2.0
djangorestframework==3.14.0